+++ tests/testthat/test-bioMarkerMod-shinytest2.R
@@ -14,8 +14,7 @@ test_that("biomarkerMod works", {
     variant = platform_variant(),
     name = "biomarkerMods-1",
     height = 968,
-    width = 1619,
-    view = interactive()
+    width = 1619
   )
 
   app$expect_screenshot()
@@ -26,10 +25,10 @@ test_that("biomarkerMod works", {
 
   app$set_inputs(
     current_studyid = "adrc00031",
-    wait_ = FALSE
+    wait_ = TRUE
   )
 
-  app$wait_for_idle(timeout = 60000)
+  # app$wait_for_idle(timeout = 60000)
 
   app$expect_screenshot()
 
@@ -37,39 +36,6 @@ test_that("biomarkerMod works", {
     current_studyid = "adrc00006",
     wait_ = FALSE
   )
-
-  app$wait_for_idle(timeout = 60000)
-
-  app$expect_screenshot()
-
-  app$stop()
-})
-
-
-test_that("biomarkerMod works", {
-  skip_on_cran()
-  skip_on_ci()
-  skip_if(is.null(getOption("panda_api_key")))
-
-  app <- AppDriver$new(
-    app_dir = biomarkerApp(
-      adrc_ptid = c("adrc00006", "adrc00031"),
-      biomarker_api = shiny::reactive(getOption("panda_api_key")),
-      testing = TRUE
-    ),
-    variant = platform_variant(),
-    name = "biomarkerMods-2",
-    height = 968,
-    width = 1619
-  )
-
-  app$set_inputs(current_studyid = "adrc00031")
-
-  app$wait_for_idle(timeout = 60000)
-
-  app$expect_screenshot()
-
-  app$set_inputs(current_studyid = "adrc00006")
 
   app$wait_for_idle(timeout = 60000)